Reorganise custom components by atomic design
