+++ web/src/components/atoms/Card/Card.stories.tsx
@@ -1,4 +1,4 @@
-import Card from './Card'
+import { Card } from './Card'
 
 export const generated = (args) => {
   return (
