+++ web/src/components/atoms/Heading/Heading.stories.tsx
@@ -1,4 +1,4 @@
-import Heading from './Heading'
+import { Heading } from './Heading'
 
 export const generated = (args) => <Heading {...args} />
