+++ web/src/components/atoms/Heading/index.ts
@@ -1,0 +1,1 @@
+export * from './Heading'
