+++ web/src/components/atoms/index.ts
@@ -1,0 +1,2 @@
+export * from './Card'
+export * from './Heading'
